Simple demo showing a console application written for the CoreCLR
The HelloWorld solution can be opened in Visual Studio,
or the project can be built from the command line using msbuild.
The build process will create a bin\Debug folder with HelloWorld.exe
which can be run like any other executable.
Please note that the managed application is in HelloWorld.dll.
HelloWorld.exe is just as simple CoreCLR host (renamed CoreConsole.exe)
which loads a manged application with the same name as the exe but with a dll extension.
Also, keep in mind that the the msbuild scripts used for this demo,
and the custom task are experimental and probably won’t work for more complex applications.