Changed string formatting APIs to use ByteSpan
I tried to also use ByteSpan for formatting other types, but only in internal implementation.
This resulted in performance regressions, as there was too many ByteSpans created (when public method enters)
and then immediately destroyed (when it terminates).
I think I would have to make ByteSpan public to see the perf wins for small datatypes,
but I am not ready to make it public yet.