Added prototype of composite formatting support
Composite formatting allow to format based on a template with one or more insertion points, e.g.
    var formatter = new StringFormatter();
    formatter.Format("{2} - Error {0}. File {1} not found.", 404, "index.html", DateTime.UtcNow);

This is very convinient, but unfortunatelly causes lots of GC heap allocations: the inseterd arguments
are often boxed and sometimes a so called "params-array" needs to be allocated.

This prototype can do composite formatting without any allocation (no boxing and no arrays), but only for
primitive types (ints, strings, chars, DateTime, TimeSpan) and only for up to 5 arguments.

Unfortunately, because not all type in the system implement IBufferFormattable, this API does not
support types that do implement it (yes, it sounds backwards buy I did not mistype that).
Composite formatting would be totally useless if it did not support built in types,
and the current design needs to trade off support for types that implement the interface to be able to support
primitive types (given the absolute goal of not boxing).
See CompositeFormattingExtensions.AppendUntyped for more details.