Added support for locales other than invatiant
For now, only UTF16 formatting is supported.
The locale data was generated from CLDR, but I have not throughly tested the data generation tool, and so it might have bugs.
Also, the tool is not yet in the repo. I will try to add it after cleanup and testing.

Lastly, I discovered a bug int the formatting code. It does not print the group separateo for 1000 (four digit numbers).
I will try to fix it soon.