Changed non-allocating web server to use native memory for request buffers
This should improve perfromance, but unfortunatelly because I am not using a real Span<byte>, the APIs
are not as safe as before. User code needs to be careful to not store Utf8Span on the heap past each request
as the native memory will be returned to the pool.