Remove support for slash options
Supporting slashes seems problematic for cross-platform usage. When
running on non-Windows platform, we'd have to disallow it anyways in
order to avoid parsing issues.

On Windows there is already enough prior-art to justify dropping support
for it. For instance, PowerShell doesn't slashes either. Also, Windows
users have started to use many ports of Unix tools, such as Git, that
also don't support slashes.

The only downside of removing support for slashes is that existing tools
that support slashes can't start to depend on this library. Until we've
anybody asking for it, it seems better to start clean.