Disallow options before commands
We currently support the notion of global options. These are options that
are defined before the first command. The problem is that their behavior
is non-intuitive.

Some tools, such as Git, have options that are valid without having to
specify a command, for example:

  $ git --version

Developers are likely to expect that they can support this by writing:

  ArgumentSyntax.Parse(args, syntax =>
  {
      syntax.DefineOption("v|version", ref version, "Gets the version");

      // syntax.DefineCommand(...);
      // ...
  }

But that's not the current behavior. The behavior of options that are
defined before the first command is that they apply to all commands,
instead of being only valid when there is no command.

The reason for that is incremental parsing. Consider this:

  $ git foo --version

When DefineOption is called, the code has to decide whether or not parse
the option. In the example above, the are no commands defined yet. In
that case, the code decides to parse the option. However, assuming there
is a foo command that defines the --version then the code wouldn't be
allowed to process the option.

It seems the only way to make this sane is by disallowing the definition
of commands when options are already defined.

The notion of global options would then have to be specified explicitly by
adding the concept of a 'blank command':

  ArgumentSyntax.Parse(args, syntax =>
  {
      syntax.DefineBlankCommand(ref command);
      syntax.DefineOption("v|version", ref version, "Gets the version");

      // syntax.DefineCommand(...);
      // ...
  }

The developer would then have the following options:

  1. Authoring a tool without commands. In that case the only usage is
     DefineOption and DefineParameter.

  2. Authoring a tool with commands. In that case all usages of
     DefineOption and DefineParameter are preceded by DefineCommand.

  3. Authoring a tool with commands AND options without commands. Same as
     (2) but additionally options can be preceded by call to
     DefineBlankCommand.