It was mostly project.json changes to have the projects depend on nuget packages as opposed to P2P dependencies.
I had System.Text.Utf8 working as well, but now it has a dependency on the custom csc compiler which I haven't figured out how to make it work for dotnet compile.

Adding unsafe option to System.Slices project.json and adding global.json

Fixing Global.json

In order to make buffers build in the cli and in VS, I needed to remove the project dependency from buffers to slices and replace it with a package dependency. Also replaced csproj with xproj

Reverting back to all csprojs, but added nuget dependencies to all project.json files. Needed to revert back to make build.cmd to build and pass in a sane manner. Building through the solution in VS was already not working, as it missed the reference to the projects outside of the src folder. This way we maintain the state of the world before and we get the CLI to build. So far I have tested Slices, Buffers and CommandLine.

Enabling dotnet compile for System.Text.Utf8