Making Reflection.CIL build in the CLI. Added project.json, removed package.json, modifed projects and added CLSCompliant attributes so that it continues to build in msbuild
Adding the right version of System.metadata.reflection to dnxcore5 framework. Modifying CilTests.cs to run inside the dnxcore50 folder.

Changing System.Reflection.Metadata.Cil test to use demo1.exe instead of mscorlib.dll as that has not been working for a while, not sure why. I will log an issue for that

Reducing the dependencies for Metadata.Cil to the bare minimum

Moving System.Reflection.Metadata.Cil from csproj to xproj and adding specific targets to xprojs that use dotnet

Adding dotnet to the tools in the build and downloading dotnet during build tool bootstrap. Also updated the dotnet targets to use the downloaded one.

Making dotnet restore happen before dotnet compile and locking lock files

Forcing msbuild to use the xproj filename as the name of the dll to be used in the test command. Ideally, it should use the name inside the project.json, but for some reason, it is not