Added an analyzer to flag stack only types on the heap.
This analyzer will flag a field if the type of the filed is a stack only type,
but the type containing the field is not stack-only.

Stack-only types are types that have an attribute with simple name equal to "StackOnlyAttribute" applied to them.
e.g.

using System;

class StackOnlyAttribute : Attribute {}

[StackOnly]
struct StackOnlyStruct {}

class SmoeClass {
    StackOnlyStruct _field;
}

The filed "_field" will be flagged.