Moving all projects under their own folder under src and the tests under their own test project under tests.
Updating project.json to reflect the new folder structure. Also, incremental building is not working, so I had to turn off parallel building in MSBuild. Lastly, I removed dotnet pack for the test projects so that they are not published to the myget feed.

Adding a corefxlab.sln file and removing the individual solutions under the individual projects.