Added ReferenceCounter
ReferenceCounter is a small datastructure to store an manipulate reference counts
It stores counts on per thread basis, which allows us to manipulate the counts without synchronization.
When the count needs to be read, it does need to be tallied up from all the threads.
We compute the tally without synchronizing with the thread-local modifications, and so
there is a possibility of a race here. But, the race is much less likely to occur than today
when Span<T> is "checked-out" from Memory<T> without incrementing the OwnedMemory<T> refcount,
and so it's an improvement. Also, we think we can eliminate the impact of the race at higher level,
i.e. in OwnedMemory<T> (at some additional memory barrier cost).

But the performance wins are very large when compared to contended Interlocked reference counting.
A quick test incrementing and decrementing the count 1M times on 10 threads shows the following results:

ReferenceCounter 206ms
Interlocked      13458ms