Unary ObjectTable for RefCounter (#1054)
* Converted to a Unary counter

Rather than storing the count, we use an entry per reference. This
means it is very quick to add a reference, we simple put the object
in the next slot of the array.

Releasing a reference, is a little harder, we scan the array looking
for the entry, once we find it we can collapse the list into a
condensed form, by moving the first element into the gap, and nudging
the first up.

The API has changed to only return a bool for whether a reference
exists or not. We don't care how many references there are. This means
we can short cut several searchs as soon as we find something.

* PR feedback and Typo