Improve race detection in ref counting
This design alters the synchronization to catch a race between
Dispose and AddReference.

By making Dispose of a OwnedMemory first invalidates the _id,
and then check for a zero reference count, and making
AddReference validate the id after increasing the count,
we guarantee that parallel calls to Dispose and Reserve at
least one will fail.