Fix race condition in ResizingThreadTableWorks.
I believe this fixes an intermittent fault on ResizingThreadTableWorks.
The race could be observed by inserting a Thread.Sleep(10) as the first
command of the thread, then the test fails consistently.

The WaitHandles are now made in sequential code, so there is no race on
adding to the threads list, with other additions, or the wait at the
end of the test.