OwnedMemory tidy and performance testing (#1116)
* Remove access to Reference Count

Do not allow direct access to the reference count, but instead query if
it is zero or not. This is a precursor to allowing more efficient
reference counting schemes.

* Removed Reserve with ReadOnlyMemory ref

This is no longer required according to the discussion in #952

* Made DisposableReservation return Spans

We need DisposableReservations to be able to return Spans, so we can
scope the access of the span.

* Split disposal from pooling in Memory test.

* Made ReferenceCounter methods static

There is no instance state for the ReferenceCounter, so make all its
methods static.

* Enable switching modes for Reservations

To enable testing different strategies for managing reservations,
introduce a static variable that determines how reservations should be
checked.

* Added performance test for Reservations

Added a simple performance test that allows a simple parameter sweep for
number of threads, objects, and amount of work.

This is not a very realistic example as it does not do the pipelining
that is the key use case currently.