Finalizer issue with OwnedNativeMemory (#1197)
* Finalizer issue with OwnedNativeMemory

The finalizer could fail if the object had got into a bad state due
to Reservation having been leaked. This occured in the test due
to trying to dispose an object while having reservations. This is
a possible fix, but it may be better to decide on how finalization
and Dispose should interact.

* Added back the tests.