Add missing PackageReference to System.Text.Primitives.Tests (#1795)
At least nuget has an excuse for missing this one - this
is an assembly that's loaded by Reflection inside the bowels
of xunit. Not that I plan to run benchmarks in the self-contained
setup but without this, you can't even skip the benchmark
tests since xunit.performance.execution.dll is the one
that contains the Discoverer class that tells you
that [Benchmark] is a "Benchmark" trait.