Constructor work & perf tweaks
Making the constructor public cuts the construction time by about a third.

Tweak UnreadSpan to inline and give back the current span if at the start.

Skip out on getting the next memory from the sequence if we're single segment.

Cuts 5-10% off of some of the perf tests.

Also change terms to Span from Segment. It was mentally taxing to juggle Span/Segment/Sequence/SequencePosition.