Initial commit of System.Reflection.TypeLoader
A TypeInfo implementation layered over System.Reflection.Metadata

Source files under FromCoreRT are taken directly from the Project N
Reflection implementation with only minimal source changes. We don't
have a mirroring scheme between that code and corefx (let alone
CoreFxLab) but they would be possible candidates for such a
mirror. So we're putting them in their own folder to highlight
that fact (and to explain why these source files look duct-taped
onto the main TypeLoader source - it's because they are.)

Almost everything in that space has to do with the
byzantine BindingFlags logic in the apis that retrieve
type members.