Added TypeLoader.GetAssemblies() api (#2454)
Returns an atomic snapshot of assemblies that have been
loaded into TypeLoader.

An internal project thinking of moving to this is using
the ReflectionOnly counterpart to this and if they don't,
it's only a matter of time before someone asks for this.

Also, dropped the precalculated MVID from the loadedAssemblyTable
storage. Loading the same assembly twice is not the common
case - no need to spend cycles looking up the MVID prematurely.