Fix handling of escaped type names in Assembly.GetType()
For some reason that I don't understand, Type.GetType()
takes the time to escape any special characters in type name
fragments before handing it to the user-supplied typeResolver delegate
(despite the fact that the typeResolve is only tasked with handling simple
non-nested types, not general type expressions, and almost no one is
likely to remember to handle those escape characters.)

However, the behavior is what it is, it is doc'd and it'd be a breaking
change to fix that. So we fix our own typeResolver to remove
the escape characters.