Add System.Web compatibility assembly
This assembly implements most the surface area of System.Web by throwing
PlatformNotSupportedException.

Some API which can be thought of as "capability" checks return default
values to indicate the API is not supported.

This enables libraries built against desktop using these parts of
System.Web to JIT successfully and execute as if they were running
outside the context of a web project.